+++ cache.h
@@ -761,7 +761,7 @@ int chmod_index_entry(struct index_state *, struct cache_entry *ce, char flip);
 int ce_same_name(const struct cache_entry *a, const struct cache_entry *b);
 void set_object_name_for_intent_to_add_entry(struct cache_entry *ce);
 int index_name_is_other(const struct index_state *, const char *, int);
-void *read_blob_data_from_index(const struct index_state *, const char *, unsigned long *);
+void *read_blob_data_from_index(const struct index_state *, const char *, size_t *);
 
 /* do stat comparison even if CE_VALID is true */
 #define CE_MATCH_IGNORE_VALID		01
@@ -1279,10 +1279,10 @@ char *xdg_cache_home(const char *filename);
 
 int git_open_cloexec(const char *name, int flags);
 #define git_open(name) git_open_cloexec(name, O_RDONLY)
-int unpack_loose_header(git_zstream *stream, unsigned char *map, unsigned long mapsize, void *buffer, unsigned long bufsiz);
-int parse_loose_header(const char *hdr, unsigned long *sizep);
+int unpack_loose_header(git_zstream *stream, unsigned char *map, size_t mapsize, void *buffer, size_t bufsiz);
+int parse_loose_header(const char *hdr, size_t *sizep);
 
-int check_object_signature(const struct object_id *oid, void *buf, unsigned long size, const char *type);
+int check_object_signature(const struct object_id *oid, void *buf, size_t size, const char *type);
 
 int finalize_object_file(const char *tmpfile, const char *filename);
 
@@ -1480,7 +1480,7 @@ int cache_name_stage_compare(const char *name1, int len1, int stage1, const char
 void *read_object_with_reference(struct repository *r,
 				 const struct object_id *oid,
 				 const char *required_type,
-				 unsigned long *size,
+				 size_t *size,
 				 struct object_id *oid_ret);
 
 struct object *repo_peel_to_type(struct repository *r,
